Two-way data binding complete and all the data gets added.

+++ app/src/main/java/com/udacity/shoestore/models/Shoe.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class Shoe(var name: String="", var size: Double=0.0, var company: String="", var description: String="") : Parcelable
+data class Shoe(var name: String="", var size: Int = 0, var company: String="", var description: String="") : Parcelable
